requirements_venv.txt
google-cloud-bigquery-storage
astronomer-cosmos[dbt-bigquery]==1.0.3
protobuf==3.20.0
apache-airflow-providers-amazon==8.26.0
apache-airflow-providers-cncf-kubernetes==8.0.1
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-google==10.17.0
google-cloud-bigquery==3.20.1